requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
cryptography==40.0.1
decorator==5.1.1
feedparser==6.0.10
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.4
pycparser==2.21
python-dotenv==1.0.0
requests==2.28.2
sgmllib3k==1.0.0
tqdm==4.65.0
urllib3==1.26.15
validators==0.20.0
weaviate-cli==2.1.3
weaviate-client==3.15.5
yarl==1.8.2
transformers~=4.28.1